@@ -58,7 +58,7 @@
 		<dependency>
 			<groupId>nl.jqno.equalsverifier</groupId>
 			<artifactId>equalsverifier</artifactId>
-			<version>2.3.2</version>
+			<version>3.15</version>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
@@ -70,14 +70,14 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.6.0</version>
 				<configuration>
-					<source>9</source>
-					<target>9</target>
+					<source>17</source>
+					<target>17</target>
 					<debug>false</debug>
 				</configuration>
 			</plugin>
 			<plugin>
 				<artifactId>maven-jar-plugin</artifactId>
-				<version>3.0.2</version>
+				<version>3.3.0</version>
 				<configuration>
 					<archive>
 						<manifest>
@@ -90,7 +90,7 @@
 			<!-- attach "bundling javadoc jar" to the package phase -->
 			<plugin>
 				<artifactId>maven-javadoc-plugin</artifactId>
-				<version>2.10.4</version>
+				<version>3.6.3</version>
 				<executions>
 					<execution>
 						<id>attach-javadoc</id>
@@ -110,7 +110,7 @@
 			<!-- attach "bundling source jar" to the package phase -->
 			<plugin>
 				<artifactId>maven-source-plugin</artifactId>
-				<version>3.0.1</version>
+				<version>3.3.0</version>
 				<executions>
 					<execution>
 						<id>attach-sources</id>
@@ -122,7 +122,7 @@
 			</plugin>
 			<plugin>
 				<artifactId>maven-surefire-plugin</artifactId>
-				<version>2.19.1</version>
+				<version>3.2.5</version>
 			</plugin>
 
 			<!-- code coverage -->
