@@ -162,7 +162,7 @@ public final class ID3v1Genres {
    * @return matching genre index or -1
    */
   public static int matchGenreDescription(String description) {
-	if (description != null && description.length() > 0) {
+	if (description != null && !description.isEmpty()) {
 	  for (int i = 0; i < ID3v1Genres.GENRES.length; i++) {
 		if (ID3v1Genres.GENRES[i].equalsIgnoreCase(description)) {
 		  return i;
