@@ -2,7 +2,7 @@ package com.mpatric.mp3agic;
 
 public final class Version {
 	private static final String VERSION;
-	private static final String URL = "http://github.com/mpatric/mp3agic";
+	private static final String URL = "https://github.com/mpatric/mp3agic";
 
 	static { // get version from JAR manifest
 		String implementationVersion = Version.class.getPackage().getImplementationVersion();
